+++ src/TMModels/Results.cs
@@ -9,7 +9,7 @@ public record Results(
 
 public record PlayerResult(
     [property: JsonPropertyName("player")] string Player,
-    [property: JsonPropertyName("totalPoints")] short TotalPoints,
+    [property: JsonPropertyName("totalPoints")] decimal TotalPoints,
     [property: JsonPropertyName("wins")] ushort Wins,
     [property: JsonPropertyName("cla")] ushort Cla,
     [property: JsonPropertyName("supplies")] ushort Supplies,
@@ -17,7 +17,7 @@ public record PlayerResult(
     [property: JsonPropertyName("minutesPerMove")] double MinutesPerMove,
     [property: JsonPropertyName("moves")] ushort Moves,
     [property: JsonPropertyName("houses")] HouseResult[] Houses,
-    [property: JsonPropertyName("penaltiesPoints")] ushort PenaltiesPoints,
+    [property: JsonPropertyName("penaltiesPoints")] decimal PenaltiesPoints,
     [property: JsonPropertyName("penaltiesDetails")] PlayerPenalty[] PenaltiesDetails
 );
 
@@ -25,7 +25,7 @@ public record HouseResult(
     [property: JsonPropertyName("game")] uint Game,
     [property: JsonPropertyName("house")] House House,
     [property: JsonPropertyName("isWinner")] bool IsWinner,
-    [property: JsonPropertyName("points")] ushort Points,
+    [property: JsonPropertyName("points")] decimal Points,
     [property: JsonPropertyName("battlePenalty")] ushort BattlePenalty,
     [property: JsonPropertyName("strongholds")] ushort Strongholds,
     [property: JsonPropertyName("castles")] ushort Castles,
@@ -38,6 +38,6 @@ public record HouseResult(
 
 public record PlayerPenalty(
     [property: JsonPropertyName("game")] uint? Game,
-    [property: JsonPropertyName("points")] ushort Points,
+    [property: JsonPropertyName("points")] decimal Points,
     [property: JsonPropertyName("details")] string Details
 );
